# docs/swagger.yaml
basePath: /api/v1
definitions:
  docsResponse.CategoryCreate201:
    properties:
      data:
        $ref: '#/definitions/haircompany-shop-rest_internal_modules_v1_category_dto.ResponseDTO'
      isSuccess:
        example: true
        type: boolean
    type: object
  docsResponse.CategoryCreate400:
    properties:
      errorCode:
        enum:
        - BAD_REQUEST
        type: string
      fields:
        items:
          $ref: '#/definitions/docsResponse.categoryErrorField'
        type: array
      isSuccess:
        example: false
        type: boolean
      message:
        example: Bad request or validation error
        type: string
    type: object
  docsResponse.CategoryDelete200:
    properties:
      data:
        $ref: '#/definitions/haircompany-shop-rest_internal_modules_v1_category_dto.ResponseDTO'
      isSuccess:
        example: true
        type: boolean
    type: object
  docsResponse.CategoryGetById200:
    properties:
      data:
        $ref: '#/definitions/haircompany-shop-rest_internal_modules_v1_category_dto.ResponseDTO'
      isSuccess:
        example: true
        type: boolean
    type: object
  docsResponse.CategoryList200:
    properties:
      data:
        items:
          $ref: '#/definitions/haircompany-shop-rest_internal_modules_v1_category_dto.ResponseDTO'
        type: array
      isSuccess:
        example: true
        type: boolean
    type: object
  docsResponse.CategoryUpdate200:
    properties:
      data:
        $ref: '#/definitions/haircompany-shop-rest_internal_modules_v1_category_dto.ResponseDTO'
      isSuccess:
        example: true
        type: boolean
    type: object
  docsResponse.CategoryUpdate400:
    properties:
      errorCode:
        enum:
        - BAD_REQUEST
        type: string
      fields:
        items:
          $ref: '#/definitions/docsResponse.categoryErrorField'
        type: array
      isSuccess:
        example: false
        type: boolean
      message:
        example: Bad request or validation error
        type: string
    type: object
  docsResponse.DashboardLogin200:
    properties:
      data:
        $ref: '#/definitions/haircompany-shop-rest_internal_modules_v1_auth_dto.ResponseDTO'
      isSuccess:
        example: true
        type: boolean
    type: object
  docsResponse.DashboardLogin400:
    properties:
      errorCode:
        enum:
        - BAD_REQUEST
        type: string
      fields:
        items:
          $ref: '#/definitions/docsResponse.authErrorField'
        type: array
      isSuccess:
        example: false
        type: boolean
      message:
        example: Bad request or validation error
        type: string
    type: object
  docsResponse.DashboardRefreshToken200:
    properties:
      data:
        $ref: '#/definitions/haircompany-shop-rest_internal_modules_v1_auth_dto.ResponseDTO'
      isSuccess:
        example: true
        type: boolean
    type: object
  docsResponse.DashboardRefreshToken400:
    properties:
      errorCode:
        enum:
        - BAD_REQUEST
        type: string
      fields:
        items:
          $ref: '#/definitions/docsResponse.authErrorField'
        type: array
      isSuccess:
        example: false
        type: boolean
      message:
        example: Bad request or validation error
        type: string
    type: object
  docsResponse.DashboardUserCreate201:
    properties:
      data:
        $ref: '#/definitions/haircompany-shop-rest_internal_modules_v1_category_dto.ResponseDTO'
      isSuccess:
        example: true
        type: boolean
    type: object
  docsResponse.DashboardUserCreate400:
    properties:
      errorCode:
        enum:
        - BAD_REQUEST
        type: string
      fields:
        items:
          $ref: '#/definitions/docsResponse.dashboardUserErrorField'
        type: array
      isSuccess:
        example: false
        type: boolean
      message:
        example: Bad request or validation error
        type: string
    type: object
...
  docsResponse.Response400:
    properties:
      errorCode:
        enum:
        - BAD_REQUEST
        type: string
      isSuccess:
        example: false
        type: boolean
      message:
        example: Bad request or validation error
        type: string
    type: object
  docsResponse.Response401:
    properties:
      errorCode:
        enum:
        - UNAUTHORIZED
        type: string
      isSuccess:
        example: false
        type: boolean
      message:
        example: Unauthorized
        type: string
    type: object
  docsResponse.Response404:
    properties:
      errorCode:
        enum:
        - NOT_FOUND
        type: string
      isSuccess:
        example: false
        type: boolean
      message:
        example: Resource not found
        type: string
    type: object
  docsResponse.Response409:
    properties:
      errorCode:
        enum:
        - HAS_LINKED_ENTITIES
        type: string
      isSuccess:
        example: false
        type: boolean
      message:
        example: Conflict
        type: string
    type: object
...
  docsResponse.Response500:
    properties:
      errorCode:
        enum:
        - SERVER_ERROR
        type: string
      isSuccess:
        example: false
        type: boolean
      message:
        example: Internal server error
        type: string
    type: object
  docsResponse.authErrorField:
    properties:
      errorCode:
        enum:
        - REQUIRED
        - INVALID_EMAIL
        - MIN_LENGTH
        - MAX_LENGTH
        - INVALID_TOKEN
        type: string
      field:
        enum:
        - email
        - password
        - refreshToken
        type: string
    type: object
  docsResponse.categoryErrorField:
    properties:
      errorCode:
        enum:
        - NOT_UNIQUE
        - NOT_FOUND
        type: string
      field:
        enum:
        - name
        - slug
        - parentId
        type: string
    type: object
  docsResponse.dashboardUserErrorField:
    properties:
      errorCode:
        enum:
        - NOT_UNIQUE
        - NOT_FOUND
        type: string
      field:
        enum:
        - name
        - slug
        - parentId
        type: string
    type: object
...
  dto.DashboardLoginDTO:
    properties:
      email:
        type: string
      password:
        maxLength: 255
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.RefreshTokenDTO:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.UpdateDTO:
    properties:
      description:
        type: string
      headerImage:
        type: string
      image:
        type: string
      isActive:
        type: boolean
      isShade:
        type: boolean
      isVisibleInMenu:
        type: boolean
      isVisibleOnMain:
        type: boolean
      name:
        maxLength: 255
        minLength: 3
        type: string
      parentId:
        type: integer
      seoDescription:
        type: string
      seoKeys:
        type: string
      seoTitle:
        type: string
      slug:
        maxLength: 255
        minLength: 3
        type: string
      sortIndex:
        minimum: 0
        type: integer
    type: object
  haircompany-shop-rest_internal_modules_v1_auth_dto.ResponseDTO:
    properties:
      refreshExpiresAt:
        type: integer
      refreshToken:
        type: string
      token:
        type: string
    type: object
  haircompany-shop-rest_internal_modules_v1_category_dto.CreateDTO:
    properties:
      description:
        type: string
      headerImage:
        maxLength: 255
        type: string
      image:
        maxLength: 255
        type: string
      isActive:
        type: boolean
      isShade:
        type: boolean
      isVisibleInMenu:
        type: boolean
      isVisibleOnMain:
        type: boolean
      name:
        maxLength: 255
        minLength: 3
        type: string
      parentId:
        description: Nullable, can be zero
        minimum: 0
        type: integer
      seoDescription:
        type: string
      seoKeys:
        type: string
      seoTitle:
        maxLength: 255
        type: string
      slug:
        maxLength: 255
        minLength: 3
        type: string
      sortIndex:
        maximum: 9999
        minimum: 0
        type: integer
    required:
    - headerImage
    - image
    - name
    - slug
    - sortIndex
    type: object
  haircompany-shop-rest_internal_modules_v1_category_dto.ResponseDTO:
    properties:
      createdAt:
        type: string
      description:
        type: string
      headerImage:
        type: string
      id:
        type: integer
      image:
        type: string
      isActive:
        type: boolean
      isShade:
        type: boolean
      isVisibleInMenu:
        type: boolean
      isVisibleOnMain:
        type: boolean
      name:
        type: string
      parentId:
        type: integer
      seoDescription:
        type: string
      seoKeys:
        type: string
      seoTitle:
        type: string
      slug:
        type: string
      sortIndex:
        type: integer
      updatedAt:
        type: string
    type: object
  haircompany-shop-rest_internal_modules_v1_dashboard_user_dto.CreateDTO:
    properties:
      email:
        type: string
      password:
        maxLength: 255
        minLength: 8
        type: string
      role:
        enum:
        - admin
        - manager
        type: string
    required:
    - email
    - password
    - role
    type: object
...
host: https://{host}:{port}
info:
  contact:
    email: x3.na.tri@gmail.com
    name: API Support
  description: REST API для магазина бренда Hair Company
  title: Hair Company Shop API
  version: "1.0"
paths:
  /api/v1/auth/dashboard/login:
    post:
      consumes:
      - application/json
      description: Authenticate dashboard user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.DashboardLoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/docsResponse.DashboardLogin200'
        "400":
          description: Bad Request or Validation Error
          schema:
            $ref: '#/definitions/docsResponse.DashboardLogin400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docsResponse.Response401'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/docsResponse.Response500'
      summary: Dashboard user login
      tags:
      - Auth
  /api/v1/auth/dashboard/refresh-token:
    post:
      consumes:
      - application/json
      description: Refresh authentication token for dashboard user
      parameters:
      - description: Refresh token
        in: body
        name: refreshToken
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/docsResponse.DashboardRefreshToken200'
        "400":
          description: Bad Request or Validation Error
          schema:
            $ref: '#/definitions/docsResponse.DashboardRefreshToken400'
        "401":
          description: Unauthorized or Invalid Token
          schema:
            $ref: '#/definitions/docsResponse.Response401'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/docsResponse.Response500'
      summary: Dashboard refresh token
      tags:
      - Auth
  /api/v1/category:
    get:
      description: Retrieve all categories
      produces:
      - application/json
      responses:
        "200":
          description: List of categories
          schema:
            $ref: '#/definitions/docsResponse.CategoryList200'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/docsResponse.Response500'
      summary: Get all categories
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Create a new category
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category to create
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/haircompany-shop-rest_internal_modules_v1_category_dto.CreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            $ref: '#/definitions/docsResponse.CategoryCreate201'
        "400":
          description: Bad Request or Validation Error
          schema:
            $ref: '#/definitions/docsResponse.CategoryCreate400'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/docsResponse.Response500'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - Category
  /api/v1/category/{id}:
    delete:
      description: Delete category by ID
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted
          schema:
            $ref: '#/definitions/docsResponse.CategoryDelete200'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/docsResponse.Response400'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/docsResponse.Response404'
        "409":
          description: Category has linked entities
          schema:
            $ref: '#/definitions/docsResponse.Response409'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/docsResponse.Response500'
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - Category
    get:
      description: Retrieve category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category found
          schema:
            $ref: '#/definitions/docsResponse.CategoryGetById200'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/docsResponse.Response400'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/docsResponse.Response404'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/docsResponse.Response500'
      summary: Get category by ID
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Update category by ID
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category update payload
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated
          schema:
            $ref: '#/definitions/docsResponse.CategoryUpdate200'
        "400":
          description: Bad request or validation error
          schema:
            $ref: '#/definitions/docsResponse.CategoryUpdate400'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/docsResponse.Response500'
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - Category
  /api/v1/dashboard-user/create:
    post:
      consumes:
      - application/json
      description: Create a new dashboard user with the provided details.
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Dashboard User Create DTO
        in: body
        name: dashboardUser
        required: true
        schema:
          $ref: '#/definitions/haircompany-shop-rest_internal_modules_v1_dashboard_user_dto.CreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Dashboard User Created
          schema:
            $ref: '#/definitions/docsResponse.DashboardUserCreate201'
        "400":
          description: Bad Request or Validation Error
          schema:
            $ref: '#/definitions/docsResponse.DashboardUserCreate400'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/docsResponse.Response500'
      security:
      - BearerAuth: []
      summary: Create a new dashboard user
      tags:
      - Dashboard User
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"